requirements.txt
absl-py==2.1.0
accelerate==0.32.1
addict==2.4.0
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
astunparse==1.6.3
basicsr==1.4.2
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cvlib==0.2.7
cycler==0.12.1
Cython==3.0.10
deepdanbooru==1.0.2
deepface==0.0.92
diffusers==0.29.2
easydict==1.13
eval_type_backport==0.2.0
facexlib==0.3.0
filelock==3.15.4
filterpy==1.4.5
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gfpgan==1.3.8
google-pasta==0.2.0
grpcio==1.65.1
gunicorn==22.0.0
h5py==3.11.0
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
imutils==0.5.4
insightface==0.7.3
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
openai-clip==1.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
prettytable==3.10.2
progressbar==2.5
protobuf==4.25.3
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
realesrgan==0.3.0
regex==2024.5.15
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
soupsieve==2.5
sympy==1.13.0
tb-nightly==2.18.0a20240718
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yapf==0.40.2
zipp==3.19.2
